Clean up the constants module
